API related fixes

+++ libraries/chain/include/hive/chain/database.hpp
@@ -785,8 +785,8 @@ namespace chain {
       transaction_id_type           _current_trx_id;
       uint32_t                      _current_block_num    = 0;
       int32_t                       _current_trx_in_block = 0;
-      uint16_t                      _current_op_in_trx    = 0;
-      uint16_t                      _current_virtual_op   = 0;
+      uint32_t                      _current_op_in_trx    = 0;
+      uint32_t                      _current_virtual_op   = 0;
 
       optional< block_id_type >     _currently_processing_block_id;
